requirements.txt
mcp[cli]
python-dotenv
langchain-google-genai>=1.0.4
langchain-mcp-adapters
langchain-core